users/urls.py
from django.urls import path 
from .views import*

from rest_framework_simplejwt.views import (
    TokenObtainPairView,
    TokenRefreshView,
    
)

urlpatterns = [
    path('login/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path("register/",RegisterUserAPIView.as_view())
]