requirements.txt
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cryptography==40.0.2
cssselect2==0.7.0
cycler==0.11.0
defusedxml==0.7.1
Django==4.2.1
django-environ==0.10.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
fonttools==4.39.4
fpdf2==2.7.4
html2pdf==0.1
html5lib==1.1
idna==3.4
kiwisolver==1.4.4
lxml==4.9.2
matplotlib==3.7.1
numpy==1.24.3
oscrypto==1.3.0
packaging==23.1
Pillow==9.5.0
pycparser==2.21
pyHanko==0.18.1
pyhanko-certvalidator==0.22.0
PyJWT==2.7.0
pyparsing==3.0.9
pypdf==3.8.1
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.4.2
reportlab==3.6.13
requests==2.30.0
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
uritools==4.0.1
urllib3==2.0.2
webencodings==0.5.1
xhtml2pdf==0.2.11
